Basic encounter with new battle system implemented. Need to display the hp values, clean up player action list, and implement allies

+++ Assets/Scripts/EnemyAction.cs
@@ -14,16 +14,16 @@ public class EnemyAction : ICloneable
 
     // Base damage (i.e. how much people want to talk about this)
     // Lower means less hp done meaning people really want to talk about it
-    public float baseDmg;
+    public int baseDmg;
 
     public int maxTargets;
 
-    public void init(int id, string title, Common.EnemyActionType type, double baseDmg, int maxTargets)
+    public void init(int id, string title, Common.EnemyActionType type, int baseDmg, int maxTargets)
     {
         this.id = id;
         this.title = title;
         this.type = type;
-        this.baseDmg = (float)baseDmg;
+        this.baseDmg = baseDmg;
         this.maxTargets = maxTargets;
     }
